--- channels_to_check.txt
@@ -1,79 +1,0 @@
-@oneclickvpnkeys
-@DirectVPN
-@Parsashonam
-@V2rayNGmat
-@fnet00
-@Outline_Vpn
-@azadi_az_inja_migzare
-@vmess_vless_v2rayng
-@v2ray_vpn_ir
-@BestV2rang
-@v2logy
-@Awlix_V2RAY
-@reality_daily
-@DeamNet_Proxy
-@vpn_go67
-@SafeNet_Server
-@mahvarehnewssat
-@sinabigo
-@FAKEOFTVC
-@customv2ray
-@VPNCUSTOMIZE
-@rxv2ray
-@MTConfig
-@MehradLearn
-@v2Line
-@v2rayNG_Matsuri
-@Helix_Servers
-@EUT_VPN
-@proxy_mtm
-@eliya_chiter0
-@melov2ray
-@servermomo
-@ARv2ray
-@vless_vmess
-@GozargahVPN
-@V2rayCollector
-@v2rayng_config_amin
-@VPNCLOP
-@DarkTeam_VPN
-@ProxyForOpeta
-@Outlinev2rayNG
-@v2_team
-@VpnFreeSec
-@freeconfigv2
-@hcv2ray
-@NIM_VPN_ir
-@Capital_NET
-@v2ray_swhil
-@XsV2ray
-@V2parsin
-@EliV2ray
-@proxyymeliii
-@V2rayCollectorDonate
-@DigiV2ray
-@free_v2rayyy
-@yaney_01
-@ShadowProxy66
-@MrV2Ray
-@V2rayNGn
-@V2pedia
-@FalconPolV2rayNG
-@CUSTOMVPNSERVER
-@MsV2ray
-@ServerNett
-@lrnbymaa
-@Proxy_PJ
-@vmessorg
-@polproxy
-@v2rayng_vpnrog
-@lightning6
-@frev2ray
-@proxy_kafee
-@Qv2rayDONATED
-@Capoit
-@PrivateVPNs
-@Cov2ray
-@kiava
-@prrofile_purple
-@nofiltering2
